LayoutTests/webanimations/combining-transform-animations-with-different-acceleration-capabilities-8.html

<div id="target" style="width: 100px; height: 100px; background-color: black;"></div>
<script src="../resources/js-test.js"></script>
<script>

const element = document.getElementById("target");
const timing = { duration: 1000, iterations: Infinity };

const shouldBecomeEqualAsync = async (actual, expected) => new Promise(resolve => shouldBecomeEqual(actual, expected, resolve));

(async () => {
    if (!window.testRunner || !window.internals) {
        debug("This test should be run in a test runner.");
        return;
    }

    testRunner.waitUntilDone();

    // Start a transform-related animation that can be accelerated.
    element.animate({ scale: [1, 2] }, timing);
    await shouldBecomeEqualAsync("internals.acceleratedAnimationsForElement(element).length", "1");

    // Start another transform-related animation that can also be accelerated.
    element.animate({ translate: ["0", "100px"] }, timing);
    await shouldBecomeEqualAsync("internals.acceleratedAnimationsForElement(element).length", "2");

    // Update the playback rate of the first animation to something other than 1.
    // This will prevent running animations in that stack to keep running accelerated.
    element.getAnimations()[0].updatePlaybackRate(0.5);
    await shouldBecomeEqualAsync("internals.acceleratedAnimationsForElement(element).length", "0");

    testRunner.notifyDone();
})();
    
</script>
